# Optimized django.utils.text.capfirst().