@smithdc1 2021-01-18
Hi @pope1ni timings look good to me. I tried a few different versions (including the marginally faster but horrid f-string version) and this looks to me like the best option. 

I'm not sure about asking `django.template.defaultfilters.capfirst()` to use this function as well. I doubt it would make sense from a pure performance perspective as function calls are particularly expensive. But if it makes the code more readable and maintainable then we should do it anyway.

Finally, I'm not convinced about bytes. Feels to me like we should let someone present a use case for this first? :shrug: