@pope1ni 2021-01-28
Hi @carltongibson 

I think this should be quite a straightforward one. Obviously calling `str()` twice is excessive, but also unnecessary if it is already of type `str`. The `isinstance(..., str)` is cheaper than creating a new object.